Добавлены getAt и setAt, исправлено множество ошибок, добавлена функция slice

+++ OperationProcessing.hpp
@@ -103,7 +103,7 @@ bool isComparisonOperation(string op) {
 	return contain({"==", "!=", ">=", "<=", ">", "<"}, op);
 }
 
-int processOperation(long long b, long long a, string op) {
+long long processOperation(long long b, long long a, string op) {
 	if (op == "+") {
 		return a + b;
 	}
@@ -139,7 +139,7 @@ int processOperation(long long b, long long a, string op) {
 	}
 }
 
-double processOperation(long double b, long double a, string op) {
+long double processOperation(long double b, long double a, string op) {
 	if (op == "+") {
 		return a + b;
 	}
